requirements.txt
django
celery[redis]
djangorestframework
requests
django-cors-headers
passlib
PyJWT
Pillow
django_redis
decorator==4.0.11
#psycopg2-binary
#psycopg2
webvtt-py
pandas
gunicorn
logmatic-python
psutil
